+++ pkg/vault/test/foundry/Hooks.t.sol
@@ -16,7 +16,6 @@ import { FixedPoint } from "@balancer-labs/v3-solidity-utils/contracts/math/Fixe
 import { PoolMock } from "../../contracts/test/PoolMock.sol";
 import { PoolFactoryMock } from "../../contracts/test/PoolFactoryMock.sol";
 import { PoolHooksMock } from "../../contracts/test/PoolHooksMock.sol";
-import { PoolConfigBits } from "../../contracts/lib/PoolConfigLib.sol";
 
 import { BaseVaultTest } from "./utils/BaseVaultTest.sol";
 
@@ -40,8 +39,8 @@ contract HooksTest is BaseVaultTest {
         (daiIdx, usdcIdx) = getSortedIndexes(address(dai), address(usdc));
 
         // Create another pool and pool factory to test onRegister
-        uint256 pauseWindowEndTime = vault.getPauseWindowEndTime();
-        uint256 bufferPeriodDuration = vault.getBufferPeriodDuration();
+        uint32 pauseWindowEndTime = vault.getPauseWindowEndTime();
+        uint32 bufferPeriodDuration = vault.getBufferPeriodDuration();
         anotherFactory = new PoolFactoryMock(IVault(vault), pauseWindowEndTime - bufferPeriodDuration);
         vm.label(address(anotherFactory), "another factory");
         anotherPool = address(new PoolMock(IVault(address(vault)), "Another Pool", "ANOTHER"));
